Faders are now stored as 'int'!
Which is actually what the console (and MIDI) deals with, so everything is easier now.
Furthermore, automatic conversion and handling of float values for legacy (TouchOSC) clients has been added, as their faders only do 'float' (lame). This also makes everything fully compatible with the new, up-and-coming YAMAHA-OSC app

+++ YAMAHA MIDI/InfoWindow.xaml.cs
@@ -71,7 +71,7 @@ namespace YAMAHA_MIDI {
 			});
 		}
 
-		void SetFadersValue (List<float> channelFaders) {
+		void SetFadersValue (List<int> channelFaders) {
 			Dispatcher.Invoke(() => {
 				for (int i = 0; i <= 15; i++) {
 					faderBars[i].Value = channelFaders[i];
